el programa funciona y ya valida si el correo no esta registrado ya tiene documentacion swagger del post:login

+++ domain/model/src/main/java/co/com/projectve/model/user/User.java
@@ -23,6 +23,6 @@ public class User {
     private String email;
     private BigDecimal baseSalary;
     private String password;
-    private byte rol;
+    private Integer rol;
 
 }
